fix: resolve pagination flash attn (#52)
* first commit

* add changes

* add changes

* add changes

* add changes

* refactor

* refactor

* refactor

* refactor

* add changes

* add random block test

* add random block test

* add random block test

* resolve compillation issues

* resolve compillation issues

* correct test

* correct test

* correct test

* correct test

* correct test

* set sample len to 128

* correct a few bugs in the test

* first commit

* casting types

* compiler issues

* compiler issues

* compiler issues

* compiler issues

* compiler issues

* first commit

* missing changes

* missing changes

* add changes

* revert back index_t type in flash.h

* revert back index_t type in flash.h

* refactor llama impl

* refactor llama impl and tests

* further changes to llama impl

* resolve further compilation issues

* remove unused code and clippy

+++ csrc/src/ffi.rs
@@ -14,21 +14,21 @@ extern "C" {
 
         is_seqlens_k_cumulative: bool,
 
-        q_batch_stride: u32,
-        k_batch_stride: u32,
-        v_batch_stride: u32,
-        o_batch_stride: u32,
-        alibi_slopes_batch_stride: u32,
+        q_batch_stride: i64,
+        k_batch_stride: i64,
+        v_batch_stride: i64,
+        o_batch_stride: i64,
+        alibi_slopes_batch_stride: i64,
 
-        q_row_stride: u32,
-        k_row_stride: u32,
-        v_row_stride: u32,
-        o_row_stride: u32,
+        q_row_stride: i64,
+        k_row_stride: i64,
+        v_row_stride: i64,
+        o_row_stride: i64,
 
-        q_head_stride: u32,
-        k_head_stride: u32,
-        v_head_stride: u32,
-        o_head_stride: u32,
+        q_head_stride: i64,
+        k_head_stride: i64,
+        v_head_stride: i64,
+        o_head_stride: i64,
 
         num_splits: u32,
 
@@ -41,7 +41,7 @@ extern "C" {
         scale_softmatx_log2: f32,
 
         block_table: *const c_int,
-        block_table_batch_stride: u32,
+        block_table_batch_stride: i64,
         page_block_size: c_int,
 
         seqused_k: *const c_int,
